requirements.txt
streamlit==1.24.1
requests==2.31.0
newspaper3k==0.2.8
lxml[html5]==4.9.3
premai==0.3.79
nltk==3.8.1
beautifulsoup4==4.12.2
cssselect==1.2.0
pillow==9.5.0
tqdm==4.65.0
feedparser==6.0.10
beautifulsoup4==4.12.2